+++ src/main/kotlin/dev/furq/holodisplays/data/display/EntityDisplay.kt
@@ -1,18 +1,22 @@
-@file:UseSerializers(Vector3fSerializer::class)
+@file:UseSerializers(Vector3fSerializer::class, QuaternionfSerializer::class)
 
 package dev.furq.holodisplays.data.display
 
+import dev.furq.holodisplays.utils.QuaternionfSerializer
 import dev.furq.holodisplays.utils.Vector3fSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
 import net.minecraft.entity.EntityPose
 import net.minecraft.entity.decoration.DisplayEntity.BillboardMode
+import org.joml.Quaternionf
 import org.joml.Vector3f
 
 @Serializable
 data class EntityDisplay(
     val id: String,
     override val rotation: Vector3f? = null,
+    override val leftRotation: Quaternionf? = null,
+    override val rightRotation: Quaternionf? = null,
     override val scale: Vector3f? = null,
     val glow: Boolean? = null,
     val pose: EntityPose? = null,
@@ -23,12 +27,14 @@ data class EntityDisplay(
     class Builder : BaseDisplay.Builder<EntityDisplay> {
         var id: String = ""
         override var rotation: Vector3f? = null
+        override var leftRotation: Quaternionf? = null
+        override var rightRotation: Quaternionf? = null
         override var scale: Vector3f? = null
         override var billboardMode: BillboardMode? = null
         var glow: Boolean? = null
         var pose: EntityPose? = null
         override var conditionalPlaceholder: String? = null
 
-        override fun build() = EntityDisplay(id, rotation, scale, glow, pose, conditionalPlaceholder)
+        override fun build() = EntityDisplay(id, rotation, leftRotation, rightRotation, scale, glow, pose, conditionalPlaceholder)
     }
 }