+++ src/main/kotlin/dev/furq/holodisplays/data/display/TextDisplay.kt
@@ -1,11 +1,13 @@
-@file:UseSerializers(Vector3fSerializer::class)
+@file:UseSerializers(Vector3fSerializer::class, QuaternionfSerializer::class)
 
 package dev.furq.holodisplays.data.display
 
+import dev.furq.holodisplays.utils.QuaternionfSerializer
 import dev.furq.holodisplays.utils.Vector3fSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
 import net.minecraft.entity.decoration.DisplayEntity.BillboardMode
+import org.joml.Quaternionf
 import org.joml.Vector3f
 
 @Serializable
@@ -18,6 +20,8 @@ data class TextDisplay(
     val seeThrough: Boolean? = null,
     val alignment: TextAlignment? = null,
     override val rotation: Vector3f? = null,
+    override val leftRotation: Quaternionf? = null,
+    override val rightRotation: Quaternionf? = null,
     override val scale: Vector3f? = null,
     override val billboardMode: BillboardMode? = null,
     override val conditionalPlaceholder: String? = null
@@ -40,13 +44,15 @@ data class TextDisplay(
         var seeThrough: Boolean? = null
         var alignment: TextAlignment? = null
         override var rotation: Vector3f? = null
+        override var leftRotation: Quaternionf? = null
+        override var rightRotation: Quaternionf? = null
         override var scale: Vector3f? = null
         override var billboardMode: BillboardMode? = null
         override var conditionalPlaceholder: String? = null
 
         override fun build() = TextDisplay(
             lines.toList(), lineWidth, backgroundColor, textOpacity,
-            shadow, seeThrough, alignment, rotation, scale, billboardMode, conditionalPlaceholder
+            shadow, seeThrough, alignment, rotation, leftRotation, rightRotation, scale, billboardMode, conditionalPlaceholder
         )
     }
 }